+++ CHANGELOG.md
@@ -45,6 +45,8 @@
 * method `Services\Main\Service::getAvailableMethods` marks as deprecated
 * method `Services\Main\Service::getAllMethods` marks as deprecated
 * method `Services\Main\Service::getMethodsByScope` marks as deprecated
+* fabric methods for `Bitrix24\SDK\Core\Credentials`
+  are [consistent](https://github.com/mesilov/bitrix24-php-sdk/issues/303): `createFromWebhook`, `createFromOAuth`, `createFromPlacementRequest`
 
 ### Bugfix